refactored duplicate method
